requirements.txt
flake8
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
numpy==1.24.1
Pillow==9.4.0
requests==2.28.1
split-folders==0.5.1
typing_extensions==4.4.0
urllib3==1.26.13
